+++ src/api/events/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+event_bp = Blueprint('events', __name__)
+
+from . import events
