+++ src/api/utils.py
@@ -1,4 +1,7 @@
 from flask import jsonify, url_for
+from flask_mail import Mail
+
+mail = Mail()
 
 
 class APIException(Exception):
