+++ src/app.py
@@ -7,9 +7,11 @@ from flask_migrate import Migrate
 from flask_swagger import swagger
 from api.utils import APIException, generate_sitemap
 from api.models import db
-from api.routes import api
 from api.admin import setup_admin
 from api.commands import setup_commands
+from api import event_bp
+from api import user_bp
+
 from flask_jwt_extended import JWTManager,create_access_token,get_jwt_identity,jwt_required
 from flask_cors import CORS
 from flask_mail import Mail
@@ -63,7 +65,9 @@ app.config['MAIL_DEFAULT_SENDER'] = 'wanderbud2024@fastmail.com'
 mail = Mail(app)
 
 # Add all endpoints form the API with a "api" prefix
-app.register_blueprint(api, url_prefix='/api')
+app.register_blueprint(event_bp, url_prefix='/api')
+app.register_blueprint(user_bp, url_prefix='/api')
+
 
 # Handle/serialize errors like a JSON object
